There, my proxy work should be up now.

+++ proxylab-handout/tiny/cgi-bin/adder.c
@@ -22,12 +22,12 @@ int main(void) {
     /* Make the response body */
     sprintf(content, "Welcome to add.com: ");
     sprintf(content, "%sTHE Internet addition portal.\r\n<p>", content);
-    sprintf(content, "%sThe answer is: %d + %d = %d\r\n<p>", 
+    sprintf(content, "%sThe answer is: %d + %d = %d\r\n<p>",
 	    content, n1, n2, n1 + n2);
     sprintf(content, "%sThanks for visiting!\r\n", content);
-  
+
     /* Generate the HTTP response */
-    printf("Content-length: %d\r\n", strlen(content));
+    printf("Content-length: %d\r\n", (int)strlen(content));
     printf("Content-type: text/html\r\n\r\n");
     printf("%s", content);
     fflush(stdout);
